pyproject.toml
[project]
name = "mongodb-analytics-agent"
version = "1.0.0"
description = "Clean MongoDB analytics system with FastMCP and LangGraph"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.127.0",
    "fastmcp>=2.14.1",
    "langchain>=0.3.0",
    "langchain-anthropic>=1.3.0",
    "langchain-google-genai>=4.1.3",
    "langchain-groq>=0.1.0",
    "langchain-mcp-adapters>=0.1.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-mongodb>=0.3.0",
    "matplotlib>=3.10.8",
    "mcp>=1.25.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
    "seaborn>=0.13.2",
    "uvicorn>=0.40.0",
]

[tool.uv.workspace]
members = [
    "src/mcp_server",
]
